refactor translate function in draw.Object

+++ test1/Draw/Object/IObject.cs
@@ -3,7 +3,7 @@
 	public interface IObject
 	{
 		void Draw(Gdk.Window window);
-		void Translate(int x, int y);
+		void Translate(double x, double y);
 		bool IsContain(Common.PointD point);
 		void ChangeColor(int r, int g, int b);
 		void ChangeColor(Cairo.Color color);
